refactor: Favorite, Management, User 도메인 안의 service 클래스를 인터페이스로 변경
