+++ server/module-core/src/main/java/com/momo/domain/management/service/ManagementService.java
@@ -1,41 +1,16 @@
 package com.momo.domain.management.service;
 
-import com.momo.domain.group.repository.ParticipantRepository;
-import com.momo.domain.post.entity.Post;
-import com.momo.domain.post.repository.PostRepository;
 import com.momo.domain.management.dto.MyPostCardResponse;
 import com.momo.domain.management.dto.ParticipatingGroupCardResponse;
 import com.momo.domain.management.dto.ParticipatingGroupCountResponse;
 import com.momo.domain.user.entity.User;
 import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class ManagementService {
+public interface ManagementService {
 
-    private final ParticipantRepository participantRepository;
-    private final PostRepository postRepository;
+    ParticipatingGroupCountResponse findParticipatingGroupCountByUser(User loginUser);
 
-    @Transactional(readOnly = true)
-    public ParticipatingGroupCountResponse findParticipatingGroupCountByUser(User loginUser) {
-        Long count = participantRepository.countAllByUser(loginUser);
-        return ParticipatingGroupCountResponse.of(count);
-    }
+    List<ParticipatingGroupCardResponse> findParticipatingGroupsByUser(User loginUser);
 
-    @Transactional(readOnly = true)
-    public List<ParticipatingGroupCardResponse> findParticipatingGroupsByUser(User loginUser) {
-        return participantRepository.findParticipatingGroupsByUser(loginUser);
-    }
-
-    @Transactional(readOnly = true)
-    public List<MyPostCardResponse> findMyPostsByUser(User loginUser, int page, int size) {
-        List<Post> posts = postRepository
-            .findAllWithGroupAndAuthorByUserOrderByCreatedDateDesc(loginUser, PageRequest.of(page, size));
-        return MyPostCardResponse.listOf(posts);
-    }
+    List<MyPostCardResponse> findMyPostsByUser(User loginUser, int page, int size);
 }
