+++ server/module-core/src/main/java/com/momo/domain/management/service/impl/ManagementServiceImpl.java
@@ -1,0 +1,42 @@
+package com.momo.domain.management.service.impl;
+
+import com.momo.domain.group.repository.ParticipantRepository;
+import com.momo.domain.management.dto.MyPostCardResponse;
+import com.momo.domain.management.dto.ParticipatingGroupCardResponse;
+import com.momo.domain.management.dto.ParticipatingGroupCountResponse;
+import com.momo.domain.management.service.ManagementService;
+import com.momo.domain.post.entity.Post;
+import com.momo.domain.post.repository.PostRepository;
+import com.momo.domain.user.entity.User;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class ManagementServiceImpl implements ManagementService {
+
+    private final ParticipantRepository participantRepository;
+    private final PostRepository postRepository;
+
+    @Transactional(readOnly = true)
+    public ParticipatingGroupCountResponse findParticipatingGroupCountByUser(User loginUser) {
+        Long count = participantRepository.countAllByUser(loginUser);
+        return ParticipatingGroupCountResponse.of(count);
+    }
+
+    @Transactional(readOnly = true)
+    public List<ParticipatingGroupCardResponse> findParticipatingGroupsByUser(User loginUser) {
+        return participantRepository.findParticipatingGroupsByUser(loginUser);
+    }
+
+    @Transactional(readOnly = true)
+    public List<MyPostCardResponse> findMyPostsByUser(User loginUser, int page, int size) {
+        List<Post> posts = postRepository
+            .findAllWithGroupAndAuthorByUserOrderByCreatedDateDesc(loginUser, PageRequest.of(page, size));
+        return MyPostCardResponse.listOf(posts);
+    }
+}
